Add new eslint rule check layer imports, public rule, fixed project, using methodology FSD, refactoring Code and fixed all problems - refs #93
