+++ config/storybook/preview.ts
@@ -1,5 +1,4 @@
 import { addDecorator } from '@storybook/react'
-import { Theme } from '../../src/app/providers/ThemeProvider'
 import {
   BrowserRouterDecorator,
 } from '../../src/shared/config/storybook/BrowserRouterDecorator/BrowserRouterDecorator'
