+++ package.json
@@ -7,7 +7,7 @@
     "start": "webpack serve --env port=3000",
     "start:vite": "vite -c ./vite.config.dev.ts",
     "start:dev": "concurrently \"npm start\" \"npm run start:dev:server\"",
-    "start:dev:vite": "concurrently \"npm start:vite\" \"npm run start:dev:server\"",
+    "start:dev:vite": "concurrently \"npm run start:vite\" \"npm run start:dev:server\"",
     "start:dev:server": "node ./json-server/index.js",
     "build:prod": "webpack --env mode=production apiUrl=https://prod-project-server-tau.vercel.app",
     "build:prod:vite": "vite build -c ./vite.config.prod.ts",
@@ -76,7 +76,7 @@
     "eslint": "^8.40.0",
     "eslint-config-standard-with-typescript": "^34.0.1",
     "eslint-import-resolver-typescript": "^3.5.5",
-    "eslint-plugin-dev-proger-eslint-plugin": "^0.0.6",
+    "eslint-plugin-dev-proger-eslint-plugin": "^0.0.7",
     "eslint-plugin-i18next": "^5.1.2",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-n": "^15.7.0",
