+++ src/app/App.tsx
@@ -1,12 +1,12 @@
 import React, { Suspense } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
+import { AppRouter } from './providers/router'
 import { classNames } from '@/shared/lib/classNames/classNames'
-import { useTheme } from '@/app/providers/ThemeProvider'
-import { AppRouter } from '@/app/providers/router'
 import { Navbar } from '@/widgets/Navbar'
 import { Sidebar } from '@/widgets/Sidebar'
 import { PageLoader } from '@/widgets/PageLoader'
 import { getUserMounted, userActions } from '@/entities/User'
+import { useTheme } from '@/shared/lib/hooks/useTheme/useTheme'
 
 export const App: React.FC = () => {
   const { theme } = useTheme()
