+++ src/app/providers/StoreProvider/config/store.ts
@@ -5,9 +5,9 @@ import {
   type ReducersMapObject,
 } from '@reduxjs/toolkit'
 import { type StateSchema } from './StateSchema'
+import { createReducerManager } from './reducerManager'
 import { counterReducer } from '@/entities/Counter'
 import { userReducer } from '@/entities/User'
-import { createReducerManager } from '@/app/providers/StoreProvider/config/reducerManager'
 import { $api } from '@/shared/api/api'
 import { saveScrollReducer } from '@/features/ScrollSave'
 import { rtkApi } from '@/shared/api/rtkApi'
