+++ src/app/providers/ThemeProvider/index.ts
@@ -1,6 +1,7 @@
 import ThemeProvider from './ui/ThemeProvider'
-import { useTheme } from './lib/useTheme'
-import { Theme } from './lib/ThemeContext'
+
+import { Theme } from '@/shared/const'
+import { useTheme } from '@/shared/lib/hooks/useTheme/useTheme'
 
 export {
   ThemeProvider,
