+++ src/entities/Article/ui/ArticleDetails/ArticleDetails.tsx
@@ -17,7 +17,7 @@ import {
 import cls from './ArticleDetails.module.scss'
 import CalendarIcon from '@/shared/assets/icons/calendar-20-20.svg'
 import EyeIcon from '@/shared/assets/icons/eye-20-20.svg'
-import { Icon } from '@/widgets/Icon'
+import { Icon } from '@/shared/ui/Icon'
 import { Avatar, AvatarSize } from '@/shared/ui/Avatar/Avatar'
 import { Text, TextSize, TextTheme } from '@/shared/ui/Text/Text'
 import { Skeleton } from '@/shared/ui/Skeleton/Skeleton'
