+++ src/entities/Article/ui/ArticleListItem/ArticleListItem.tsx
@@ -11,8 +11,8 @@ import cls from './ArticleListItem.module.scss'
 import { classNames } from '@/shared/lib/classNames/classNames'
 import { Text } from '@/shared/ui/Text/Text'
 import EyeIcon from '@/shared/assets/icons/eye-20-20.svg'
-import { Icon } from '@/widgets/Icon'
-import { Card } from '@/widgets/Card'
+import { Icon } from '@/shared/ui/Icon'
+import { Card } from '@/shared/ui/Card'
 import { Avatar } from '@/shared/ui/Avatar/Avatar'
 import { Button, ThemeButton } from '@/shared/ui/Button/Button'
 import { AppLink } from '@/shared/ui/AppLink/AppLink'
