+++ src/entities/Notification/ui/NotificationItem/NotificationItem.tsx
@@ -3,7 +3,7 @@ import { type Notification } from '../../model/types/notification'
 import cls from './NotificationItem.module.scss'
 import { classNames } from '@/shared/lib/classNames/classNames'
 import { TextSize, Text } from '@/shared/ui/Text/Text'
-import { Card, CardTheme } from '@/widgets/Card'
+import { Card, CardTheme } from '@/shared/ui/Card'
 import { AppLink } from '@/shared/ui/AppLink/AppLink'
 
 interface NotificationItemProps {
