+++ src/entities/Profile/ui/ProfileCard/ProfileCard.stories.tsx
@@ -2,7 +2,6 @@ import type { ComponentStory, ComponentMeta } from '@storybook/react'
 
 import { ProfileCard } from './ProfileCard'
 import { ThemeDecorator } from '@/shared/config/storybook/ThemeDecorator/ThemeDecorator'
-import { Theme } from '@/app/providers/ThemeProvider'
 import { StoreDecorator } from '@/shared/config/storybook/StoreDecorator/StoreDecorator'
 import { Currency } from '@/entities/Currency'
 import { Country } from '@/entities/Country'
