+++ src/entities/Rating/ui/RatingCard/RatingCard.tsx
@@ -5,7 +5,7 @@ import cls from './RatingCard.module.scss'
 import { classNames } from '@/shared/lib/classNames/classNames'
 import { Text } from '@/shared/ui/Text/Text'
 import { Input } from '@/shared/ui/Input/Input'
-import { Card } from '@/widgets/Card'
+import { Card } from '@/shared/ui/Card'
 import { HStack, VStack } from '@/shared/ui/Stack'
 import { StarRating } from '@/shared/ui/StarRating/StarRating'
 import { Modal } from '@/shared/ui/Modal/Modal'
