+++ src/features/LangSwitcher/ui/LangSwitcher.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { LangSwitcher } from './LangSwitcher'
 import { ThemeDecorator } from '@/shared/config/storybook/ThemeDecorator/ThemeDecorator'
-import { Theme } from '@/app/providers/ThemeProvider'
 
 const meta: Meta<typeof LangSwitcher> = {
   title: 'widgets/LangSwitcher',
