+++ src/features/ThemeSwitcher/ui/ThemeSwitcher.stories.tsx
@@ -1,7 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { ThemeSwitcher } from './ThemeSwitcher'
 import { ThemeDecorator } from '@/shared/config/storybook/ThemeDecorator/ThemeDecorator'
-import { Theme } from '@/app/providers/ThemeProvider'
+
+import { Theme } from '@/shared/const'
 
 const meta: Meta<typeof ThemeSwitcher> = {
   title: 'widgets/ThemeSwitcher',
