+++ src/features/ThemeSwitcher/ui/ThemeSwitcher.tsx
@@ -1,9 +1,10 @@
 import React, { memo } from 'react'
 import { classNames } from '@/shared/lib/classNames/classNames'
-import { Theme, useTheme } from '@/app/providers/ThemeProvider'
 import { Button } from '@/shared/ui/Button/Button'
 import LightIcon from '@/shared/assets/icons/theme-light.svg'
 import DarkIcon from '@/shared/assets/icons/theme-dark.svg'
+import { Theme } from '@/shared/const'
+import { useTheme } from '@/shared/lib/hooks/useTheme/useTheme'
 
 interface ThemeSwitcherProps {
   className?: string
