+++ src/features/notificationButton/ui/NotificationButton.tsx
@@ -4,7 +4,7 @@ import cls from './NotificationButton.module.scss'
 import { Popover } from '@/shared/ui/Popups'
 import { Button, ThemeButton } from '@/shared/ui/Button/Button'
 import NotificationIcon from '@/shared/assets/icons/norification.svg'
-import { Icon } from '@/widgets/Icon'
+import { Icon } from '@/shared/ui/Icon'
 import { NotificationList } from '@/entities/Notification'
 import { Drawer } from '@/shared/ui/Drawer/Drawer'
