+++ src/index.tsx
@@ -25,3 +25,4 @@ root.render(
     </StoreProvider>
   </BrowserRouter>
 )
+export { useTheme } from '@/shared/lib/hooks/useTheme/useTheme'
