+++ src/shared/config/storybook/StyleDecorator/StyleDecorator.tsx
@@ -1,4 +1,5 @@
 import { type Story } from '@storybook/react'
+// eslint-disable-next-line dev-proger-eslint-plugin/layer-imports
 import '@/app/styles/index.scss'
 
 export const StyleDecorator = (StoryComponent: Story) => (
