+++ src/shared/const/index.ts
@@ -25,3 +25,4 @@ export const RoutePath: Record<AppRoutes, string> = {
 
   [AppRoutes.NOT_FOUND]: '*',
 }
+export { Theme } from '@/shared/const/theme'
