+++ src/shared/const/localstorage.ts
@@ -1,1 +1,2 @@
 export const USER_LOCAL_STORAGE_KEY: string = 'user'
+export const LOCAL_STORAGE_THEME_KEY = 'theme'
