+++ src/shared/const/theme.ts
@@ -1,0 +1,5 @@
+export enum Theme {
+  LIGHT = 'app_light_theme',
+  DARK = 'app_dark_theme',
+  RED = 'app_red_theme'
+}
