+++ src/shared/lib/hooks/useTheme/useTheme.ts
@@ -1,5 +1,7 @@
 import React from 'react'
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from './ThemeContext'
+import { ThemeContext } from '../../context/ThemeContext'
+import { Theme } from '@/shared/const'
+import { LOCAL_STORAGE_THEME_KEY } from '@/shared/const/localstorage'
 
 interface UseThemeResult {
   toggleTheme: () => void
