+++ src/shared/ui/Card/index.ts
