+++ src/shared/ui/Card/ui/Card.module.scss
