+++ src/shared/ui/Code/Code.tsx
@@ -1,9 +1,9 @@
 import React, { memo } from 'react'
 import { Button } from '../Button/Button'
+import { Icon } from '../Icon'
 import cls from './Code.module.scss'
 import { classNames } from '@/shared/lib/classNames/classNames'
 import CopyIcon from '@/shared/assets/icons/copy-20-20.svg'
-import { Icon } from '@/widgets/Icon'
 
 interface CodeProps {
   className?: string
