+++ src/shared/ui/Drawer/Drawer.tsx
@@ -3,7 +3,6 @@ import { Portal } from '../Portal/Portal'
 import { Overlay } from '../Overlay/Overlay'
 import cls from './Drawer.module.scss'
 import { classNames, type Mods } from '@/shared/lib/classNames/classNames'
-import { useTheme } from '@/app/providers/ThemeProvider'
 import { AnimationProvider, useAnimationLibs } from '@/shared/lib/components/AnimationProvider'
 
 interface DrawerProps {
