+++ src/shared/ui/Icon/ui/Icon.module.scss
