+++ src/shared/ui/Icon/ui/Icon.tsx
