+++ src/shared/ui/Loader/Loader.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { ThemeDecorator } from '../../config/storybook/ThemeDecorator/ThemeDecorator'
 import { Loader } from './Loader'
-import { Theme } from '@/app/providers/ThemeProvider'
 
 const meta: Meta<typeof Loader> = {
   title: 'shared/Loader',
