+++ src/shared/ui/Modal/Modal.stories.tsx
@@ -1,7 +1,6 @@
 import type { ComponentStory, Meta } from '@storybook/react'
 import { ThemeDecorator } from '../../config/storybook/ThemeDecorator/ThemeDecorator'
 import { Modal } from './Modal'
-import { Theme } from '@/app/providers/ThemeProvider'
 
 const meta: Meta<typeof Modal> = {
   title: 'shared/Modal',
