+++ src/shared/ui/StarRating/StarRating.tsx
@@ -1,8 +1,8 @@
 import React, { memo } from 'react'
+import { Icon } from '../Icon'
 import cls from './StarRating.module.scss'
 import { classNames } from '@/shared/lib/classNames/classNames'
 import StarIcon from '@/shared/assets/icons/star.svg'
-import { Icon } from '@/widgets/Icon'
 
 interface StarRatingProps {
   className?: string
