+++ src/shared/ui/Tabs/Tabs.tsx
@@ -1,7 +1,7 @@
 import React, { memo } from 'react'
+import { Card, CardTheme } from '../Card'
 import cls from './Tabs.module.scss'
 import { classNames } from '@/shared/lib/classNames/classNames'
-import { Card, CardTheme } from '@/widgets/Card'
 
 export interface TabItem {
   value: string
