+++ src/widgets/PageError/ui/PageError.stories.tsx
@@ -1,7 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { PageError } from './PageError'
 import { ThemeDecorator } from '@/shared/config/storybook/ThemeDecorator/ThemeDecorator'
-import { Theme } from '@/app/providers/ThemeProvider'
+
+import { Theme } from '@/shared/const'
 
 const meta: Meta<typeof PageError> = {
   title: 'widgets/PageError',
