+++ src/widgets/PageLoader/ui/PageLoader.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import { PageLoader } from './PageLoader'
 import { ThemeDecorator } from '@/shared/config/storybook/ThemeDecorator/ThemeDecorator'
-import { Theme } from '@/app/providers/ThemeProvider'
 
 const meta: Meta<typeof PageLoader> = {
   title: 'widgets/PageLoader',
