+++ src/widgets/Sidebar/ui/Sidebar/Sidebar.tsx
@@ -5,8 +5,8 @@ import { getSidebarItems } from '../../model/selectors/getSidebarItems'
 import { type SidebarItemType } from '../../model/types/sidebar'
 import cls from './Sidebar.module.scss'
 import { classNames } from '@/shared/lib/classNames/classNames'
-import { ThemeSwitcher } from '@/widgets/ThemeSwitcher'
-import { LangSwitcher } from '@/widgets/LangSwitcher'
+import { ThemeSwitcher } from '@/features/ThemeSwitcher'
+import { LangSwitcher } from '@/features/LangSwitcher'
 import { Button, SizeButton, ThemeButton } from '@/shared/ui/Button/Button'
 import { VStack } from '@/shared/ui/Stack'
